backend/requirements.txt
fastapi
uvicorn
plaid-python
python-dotenv
apscheduler
sqlalchemy[asyncio]
asyncpg
alembic
cryptography
